Overhauled get functions
I had to change how we implimentaed all the get functions, as the constructor methods in the Account, Transaction, classes require a reference to their higher level counterpart such as the relation between a User class having an ArrayList of Accounts, and the Account needing to know what User it lies under. Because of this we need to pass Objects, not String IDs to each of the Get functions. Plus the only function going to call these classes will have to be the User. The rest of the functions may need to be re-evaluated