Added toJSON functions to all Models
These functions will be used to send data from the java servlets to the javascript based front-end. Its easier to handle JSON returns on the front-end as an API, than formatting HTML. As any changes in the HTML requires changes to the front-end aswell. With JSON, we can handle the same data with javascript which is faster, and easier to handle/debug