Fixed minor API fixes for WithdrawServlet
I had to change billType to memo on Rudy's servlet code, and add the checkers again. Otherwise the code is finished and works. I also experimented with the file deletion service, but THAT IS NOT AN OPTION currently. As you can't even download the file after it appears on the page. But thats ok, security isn't that big of an issue ATM.